Further reorganize `trace` serialization module into smaller focused files
## Refactor: reorganize trace serialization module into smaller, focused files

Changes:
- Extract `TRValue` hierarchy and related utilities into `TraceRecorderValue.kt`.
- Extract `TraceMetaInfo`, `TraceCollectingStrategy`, `TraceContextSavedState`, `DataBlock`, and file I/O helpers into their own files.
- Rename files for clarity:
    - `SerializationBase` → `TraceWriter`
    - `Deserialization` → `LazyTraceReader`
    - `SerializationPack` → `TracePacking`
    - `SerializationStream` → `FileStreamingSerialization`
    - `SerializationInmemory` → `InMemorySerialization`,
    - `SerializationNull` → `NullSerialization`
    - `SerializationCommon` → `Common`

- Merge `AccessLocation` serialization logic into `TraceWriter`.

- Clean up unused imports.

---------
