+++ trace/src/main/org/jetbrains/lincheck/trace/serialization/InMemorySerialization.kt
