+++ trace/src/main/org/jetbrains/lincheck/trace/serialization/NullSerialization.kt
