+++ trace/src/main/org/jetbrains/lincheck/trace/serialization/TraceMetaInfo.kt
@@ -1,63 +1,10 @@
-/*
- * Lincheck
- *
- * Copyright (C) 2019 - 2025 JetBrains s.r.o.
- *
- * This Source Code Form is subject to the terms of the
- * Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
- * with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
 package org.jetbrains.lincheck.trace.serialization
 
-import org.jetbrains.lincheck.descriptors.*
-import org.jetbrains.lincheck.trace.DiffStatus
-import org.jetbrains.lincheck.trace.TraceContext
 import org.jetbrains.lincheck.util.Logger
 import java.io.BufferedReader
-import java.io.DataInput
-import java.io.DataOutput
-import java.io.File
-import java.io.IOException
 import java.io.InputStream
 import java.io.InputStreamReader
-import java.io.OutputStream
 import java.lang.management.ManagementFactory
-
-/**
- * This file contains data structures and extensions for DataInput and DataOutput to save "primitive" trace data.
- *
- * This code is not split into serialization and deserialization parts to keep pairs of functions
- * together, so it is easy to verify that they are true pairs and use the same formats.
- *
- * These functions are for low-level object representations, without kinds and ids.
- */
-
-internal const val TRACE_MAGIC : Long = 0x706e547124ee5f70L
-internal const val INDEX_MAGIC : Long = TRACE_MAGIC.inv()
-internal const val TRACE_VERSION : Long = 21
-
-// This suffix is not enforced, but IDEA plugin rely on it
-const val DATA_FILENAME_EXT = "trace"
-// These are enforced
-const val INDEX_FILENAME_EXT = "idx"
-const val PACK_FILENAME_EXT = "packedtrace"
-internal const val ID_MAP_FILENAME_EXT = "idmap"
-internal const val THREAD_MAP_FILENAME_EXT = "threadmap"
-
-internal const val PACKED_DATA_ITEM_NAME = "trace.$DATA_FILENAME_EXT"
-internal const val PACKED_INDEX_ITEM_NAME = "trace.$INDEX_FILENAME_EXT"
-internal const val PACKED_META_ITEM_NAME = "info.txt"
-internal const val PACKED_DIFF_MARKER_ITEM_NAME = ".diff"
-internal const val PACKED_ID_MAP_ITEM_NAME = "diff.${ID_MAP_FILENAME_EXT}"
-internal const val PACKED_THREAD_MAP_ITEM_NAME = "diff.${THREAD_MAP_FILENAME_EXT}"
-internal const val PACKED_LEFT_META_ITEM_NAME = "info.left.txt"
-internal const val PACKED_RIGHT_META_ITEM_NAME = "info.right.txt"
-
-internal const val BLOCK_HEADER_SIZE: Int = Byte.SIZE_BYTES + Int.SIZE_BYTES
-internal const val BLOCK_FOOTER_SIZE: Int = Byte.SIZE_BYTES
-
-internal const val INDEX_CELL_SIZE: Int = Byte.SIZE_BYTES + Int.SIZE_BYTES + Long.SIZE_BYTES * 2
 
 /**
  * Information about conditions in which trace was collected.
@@ -315,258 +262,4 @@ data class TraceMetaInfo private constructor(
             }
         }
     }
-}
-
-internal enum class ObjectKind {
-    THREAD_NAME,
-    CLASS_DESCRIPTOR,
-    METHOD_DESCRIPTOR,
-    FIELD_DESCRIPTOR,
-    VARIABLE_DESCRIPTOR,
-    STRING,
-    ACCESS_PATH,
-    CODE_LOCATION,
-    TRACEPOINT,
-    TRACEPOINT_FOOTER,
-    BLOCK_START,
-    BLOCK_END,
-    EOF
-}
-
-internal enum class CodeLocationKind {
-    LINE,
-    ACCESS,
-    METHOD_CALL
-}
-
-internal val CodeLocation.kind: CodeLocationKind get() = when (this) {
-    is LineCodeLocation -> CodeLocationKind.LINE
-    is AccessCodeLocation -> CodeLocationKind.ACCESS
-    is MethodCallCodeLocation -> CodeLocationKind.METHOD_CALL
-}
-
-internal enum class AccessLocationKind {
-    LOCAL_VARIABLE,
-    STATIC_FIELD,
-    OBJECT_FIELD,
-    ARRAY_ELEMENT_BY_INDEX,
-    ARRAY_ELEMENT_BY_NAME,
-    UNKNOWN
-}
-
-internal fun DataOutput.writeType(value: Types.Type) {
-    when (value) {
-        is Types.ArrayType -> {
-            writeByte(0)
-            writeType(value.elementType)
-        }
-        is Types.BooleanType -> writeByte(1)
-        is Types.ByteType -> writeByte(2)
-        is Types.CharType -> writeByte(3)
-        is Types.DoubleType -> writeByte(4)
-        is Types.FloatType -> writeByte(5)
-        is Types.IntType -> writeByte(6)
-        is Types.LongType -> writeByte(7)
-        is Types.ObjectType -> {
-            writeByte(8)
-            writeUTF(value.className)
-        }
-        is Types.ShortType ->writeByte(9)
-        is Types.VoidType -> writeByte(10)
-    }
-}
-
-internal fun DataInput.readType(): Types.Type {
-    val type = readByte()
-    return when (type.toInt()) {
-        0 -> Types.ArrayType(readType())
-        1 -> Types.BOOLEAN_TYPE
-        2 -> Types.BYTE_TYPE
-        3 -> Types.CHAR_TYPE
-        4 -> Types.DOUBLE_TYPE
-        5 -> Types.FLOAT_TYPE
-        6 -> Types.INT_TYPE
-        7 -> Types.LONG_TYPE
-        8 -> Types.ObjectType(readUTF())
-        9 -> Types.SHORT_TYPE
-        10 -> Types.VOID_TYPE
-        else -> error("Unknown Type id $type")
-    }
-}
-
-internal fun DataOutput.writeMethodType(value: Types.MethodType) {
-    writeType(value.returnType)
-    writeInt(value.argumentTypes.size)
-    value.argumentTypes.forEach {
-        writeType(it)
-    }
-}
-
-internal fun DataInput.readMethodType(): Types.MethodType {
-    val returnType = readType()
-    val count = readInt()
-    val argumentTypes = mutableListOf<Types.Type>()
-    repeat(count) {
-        argumentTypes.add(readType())
-    }
-    return Types.MethodType(argumentTypes, returnType)
-}
-
-internal fun DataOutput.writeClassDescriptor(value: ClassDescriptor) {
-    writeUTF(value.name)
-}
-
-internal fun DataInput.readClassDescriptor(context: TraceContext): ClassDescriptor {
-    return ClassDescriptor(context, readUTF())
-}
-
-internal fun DataOutput.writeMethodDescriptor(value: MethodDescriptor) {
-    writeInt(value.classId)
-    writeMethodSignature(value.methodSignature)
-}
-
-internal fun DataInput.readMethodDescriptor(context: TraceContext): MethodDescriptor {
-    return MethodDescriptor(context,readInt(), readMethodSignature())
-}
-
-internal fun DataOutput.writeMethodSignature(value: MethodSignature) {
-    writeUTF(value.name)
-    writeMethodType(value.methodType)
-}
-
-internal fun DataInput.readMethodSignature(): MethodSignature {
-    return MethodSignature(readUTF(), readMethodType())
-}
-
-internal fun DataOutput.writeFieldDescriptor(value: FieldDescriptor) {
-    writeInt(value.classId)
-    writeUTF(value.fieldName)
-    writeType(value.type)
-    writeBoolean(value.isStatic)
-    writeBoolean(value.isFinal)
-    writeBoolean(value.isVolatile)
-}
-
-internal fun DataInput.readFieldDescriptor(context: TraceContext): FieldDescriptor {
-    return FieldDescriptor(
-        context = context,
-        classId = readInt(),
-        fieldName = readUTF(),
-        type = readType(),
-        fieldKind = FieldKind.fromIsStatic(isStatic = readBoolean()),
-        isFinal = readBoolean(),
-        isVolatile = readBoolean(),
-    )
-}
-
-internal fun DataOutput.writeVariableDescriptor(value: VariableDescriptor) {
-    writeUTF(value.name)
-    writeType(value.type)
-}
-
-internal fun DataInput.readVariableDescriptor(context: TraceContext): VariableDescriptor {
-    return VariableDescriptor(context, readUTF(), readType())
-}
-
-internal fun DataInput.readAccessLocation(context: TraceContext): AccessLocation {
-    val type = readByte().toInt()
-    val kind = if (type >= 0 && type < AccessLocationKind.entries.size) AccessLocationKind.entries[type]
-               else AccessLocationKind.UNKNOWN
-    when (kind) {
-        AccessLocationKind.LOCAL_VARIABLE -> {
-            val variableDescriptorId = readInt()
-            val vd = context.variablePool[variableDescriptorId]
-            return LocalVariableAccessLocation(vd)
-        }
-        AccessLocationKind.STATIC_FIELD -> {
-            val fieldDescriptorId = readInt()
-            val fd = context.fieldPool[fieldDescriptorId]
-            return StaticFieldAccessLocation(fd)
-        }
-        AccessLocationKind.OBJECT_FIELD -> {
-            val fieldDescriptorId = readInt()
-            val fd = context.fieldPool[fieldDescriptorId]
-            return ObjectFieldAccessLocation(fd)
-        }
-        AccessLocationKind.ARRAY_ELEMENT_BY_INDEX -> {
-            val index = readInt()
-            return ArrayElementByIndexAccessLocation(index)
-        }
-        AccessLocationKind.ARRAY_ELEMENT_BY_NAME -> {
-            val accessPathId = readInt()
-            val accessPath = context.getAccessPath(accessPathId)
-            return ArrayElementByNameAccessLocation(accessPath)
-        }
-        AccessLocationKind.UNKNOWN -> error("Unknown access location kind: read byte is $type, while expected one of: ${AccessLocationKind.entries.map { it.ordinal }.joinToString(", ")}")
-    }
-}
-
-internal fun DataOutput.writeKind(value: ObjectKind): Unit = writeByte(value.ordinal)
-
-internal fun DataOutput.writeCodeLocationKind(value: CodeLocationKind): Unit = writeByte(value.ordinal)
-
-internal fun DataOutput.writeAccessLocationKind(value: AccessLocationKind): Unit = writeByte(value.ordinal)
-
-internal fun DataInput.readKind(): ObjectKind {
-    val ordinal = readByte()
-    val values = ObjectKind.entries
-    if (ordinal < 0 || ordinal > values.size) {
-        throw IOException("Cannot read ObjectKind: unknown ordinal $ordinal")
-    }
-    return values[ordinal.toInt()]
-}
-
-internal fun DataInput.readCodeLocationKind(): CodeLocationKind {
-    val ordinal = readByte()
-    val values = CodeLocationKind.entries
-    if (ordinal < 0 || ordinal > values.size) {
-        throw IOException("Cannot read CodeLocationKind: unknown ordinal $ordinal")
-    }
-    return values[ordinal.toInt()]
-}
-
-internal fun DataOutput.writeDiffStatus(value: DiffStatus?): Unit = writeByte(value?.ordinal ?: - 1)
-
-internal fun DataInput.readDiffStatus(): DiffStatus? {
-    val ordinal = readByte().toInt()
-    if (ordinal == -1) return null
-
-    val values = DiffStatus.entries
-    if (ordinal < 0 || ordinal > values.size) {
-        throw IOException("Cannot read DiffStatus: unknown ordinal $ordinal")
-    }
-    return values[ordinal]
-}
-
-data class TraceOutputStreams(
-    val dataStream: OutputStream,
-    val indexStream: OutputStream,
-)
-
-/**
- * This function is used to create a pair of files for data and index in once.
- *
- * It uses [openNewFile] to open files and add [OUTPUT_BUFFER_SIZE] buffering
- * around "naked" stream.
- *
- * Data stream is created with passed name (without any additional extension),
- * and index file is named by adding [INDEX_FILENAME_EXT] extension to base name.
- */
-internal fun openNewStandardDataAndIndex(baseFileName: String): TraceOutputStreams {
-    val dataStream = openNewFile(baseFileName).buffered(OUTPUT_BUFFER_SIZE)
-    val indexStream = openNewFile("$baseFileName.$INDEX_FILENAME_EXT").buffered(OUTPUT_BUFFER_SIZE)
-    return TraceOutputStreams(dataStream, indexStream)
-}
-
-internal fun openNewFile(name: String): OutputStream {
-    val f = File(name)
-    f.parentFile?.mkdirs()
-    f.createNewFile()
-    return f.outputStream()
-}
-
-internal fun openExistingFile(name: String?): InputStream? {
-    val f = File(name)
-    if (!f.exists()) return null
-    return f.inputStream()
 }
